add compass page functionality
The compass page serves as base for the part of the game where the
user has to actually walk. Compass gets the user location and
generates a random destiny. The user gets feedback of distance and
direction (green arrow) of the destiny. On location or orientation
changes this feedback is updated. When the user reaches the destiny
the green arrow changes color to blue.